backend/requirements.txt
flask
flask-cors
keras
tensorflow
opencv-python
numpy
scipy
pandas
matplotlib
imageio
openai
elevenlabs
python-dotenv